Extract chart options dialog into its own file
